# Compare errors when fuzzing