--- a/Makefile
+++ b/Makefile
@@ -11,6 +11,7 @@ go-fuzz-corpus := ${GOPATH}/src/github.com/dvyukov/go-fuzz-corpus
 go-fuzz-dep := ${GOPATH}/src/github.com/dvyukov/go-fuzz/go-fuzz-dep
 
 test:
+	go test -v -cover ./ascii
 	go test -v -cover ./json
 	go test -v -cover ./iso8601
 	go run ./json/bugs/issue11/main.go