--- a/json/codec.go
+++ b/json/codec.go
@@ -895,18 +895,26 @@ func unexpectedEOF(b []byte) error {
 	return syntaxError(b, "unexpected end of JSON input")
 }
 
-func syntaxError(b []byte, msg string, args ...interface{}) error {
-	e := new(SyntaxError)
-	t := reflect.TypeOf(e).Elem()
-	p := unsafe.Pointer(e)
+var syntaxErrorMsgOffset = ^uintptr(0)
 
+func init() {
+	t := reflect.TypeOf(SyntaxError{})
 	for i, n := 0, t.NumField(); i < n; i++ {
 		if f := t.Field(i); f.Type.Kind() == reflect.String {
-			// Hack to set the unexported `msg` field.
-			*(*string)(unsafe.Pointer(uintptr(p) + f.Offset)) = "json: " + fmt.Sprintf(msg, args...) + ": " + prefix(b)
+			syntaxErrorMsgOffset = f.Offset
 		}
 	}
+}
 
+func syntaxError(b []byte, msg string, args ...interface{}) error {
+	e := new(SyntaxError)
+	i := syntaxErrorMsgOffset
+	if i != ^uintptr(0) {
+		s := "json: " + fmt.Sprintf(msg, args...) + ": " + prefix(b)
+		p := unsafe.Pointer(e)
+		// Hack to set the unexported `msg` field.
+		*(*string)(unsafe.Pointer(uintptr(p) + i)) = s
+	}
 	return e
 }