Ref: https://github.com/segmentio/encoding/issues/3

Running this briefly locally reported some errors. For example, given the input `["\b"]` (backspace char in string), `encoding/json` reports an error, but `segmentio/encoding/json` accepts it.