@pelletier 2019-12-04
Here is a broader sample of crashes: https://gist.github.com/pelletier/5b4079b63204651e26bd32438e01cce1