@achille-roussel 2019-12-05
@pelletier I fixed all the issues that were detected by the fuzz test, the package seems to always be returning errors of the same type than `encoding/json` now.

Some error strings are still different because we use more detailed error messages in some cases and I didn't want to lose that.