@pelletier 2019-12-05
Great! Thanks for taking care of this. Changes seem to have some impact on the basic Unmarshal tests (this branch vs master):

```
benchcmp segmentio-encoding-json-master.txt segmentio-encoding-json.txt
benchmark                                  old ns/op     new ns/op     delta
BenchmarkMarshal/*json.codeResponse2       5296312       5271759       -0.46%
BenchmarkUnmarshal/*json.codeResponse2     8606423       9043968       +5.08%

benchmark                                  old MB/s     new MB/s     speedup
BenchmarkMarshal/*json.codeResponse2       366.38       368.09       1.00x
BenchmarkUnmarshal/*json.codeResponse2     225.47       214.56       0.95x

benchmark                                  old allocs     new allocs     delta
BenchmarkMarshal/*json.codeResponse2       0              0              +0.00%
BenchmarkUnmarshal/*json.codeResponse2     35             35             +0.00%

benchmark                                  old bytes     new bytes     delta
BenchmarkMarshal/*json.codeResponse2       0             0             +0.00%
BenchmarkUnmarshal/*json.codeResponse2     7705          7909          +2.65%
```

Still, down to merge it anyway, as we promise to error when encoding/json does.