@achille-roussel 2019-12-05
Yes I assumed it would, I think it’s due to the check that characters in a string are all printable ascii, I’m working on optimizing this. If I don’t have something working well by end of day let’s just go with the current version.